% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineModel_MS.R
\name{defineModel_MS}
\alias{defineModel_MS}
\title{defineModel_MS}
\usage{
defineModel_MS(inclPhenology = TRUE, inclPanTrap = TRUE)
}
\value{
a set of code
}
\description{
defineModel_MS
}
\details{
Defines the Nimble model.
}
