% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineModel_SS.R
\name{defineModel_SS}
\alias{defineModel_SS}
\title{defineModel_SS}
\usage{
defineModel_SS(inclPhenology = TRUE, inclPanTrap = TRUE)
}
\value{
a set of code
}
\description{
defineModel_SS
}
\details{
Defines the Nimble model for one species.
}
