% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{formatData}
\alias{formatData}
\title{formatData}
\usage{
formatData(inData)
}
\arguments{
\item{inData}{A dataset produced by the simulations}
}
\value{
list of two data frames
}
\description{
formatData
}
\details{
Formats data ready for Nimble
}
