% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatMetadata.R
\name{formatMetadata}
\alias{formatMetadata}
\title{Format Metadata}
\usage{
formatMetadata(
  indat,
  incl2ndTransect = TRUE,
  inclPanTrap = TRUE,
  trueTrend = NULL,
  spPool = NULL
)
}
\arguments{
\item{indat}{A dataset produced by the simulations}

\item{incl2ndTransect}{should the model include data from the second transect walk?}

\item{inclPanTrap}{should the model include pan trap data?}
}
\value{
A one line dataframe
}
\description{
Format Metadata
}
