% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessSimDatFile.R
\name{ProcessSimDatFile}
\alias{ProcessSimDatFile}
\title{ProcessSimDatFile}
\usage{
ProcessSimDatFile(
  filename,
  inPath = ".",
  outPath = NULL,
  useNimble = TRUE,
  inclPhenology = TRUE,
  incl2ndTransect = TRUE,
  inclPanTrap = TRUE,
  inclStateRE = TRUE,
  multiSp = FALSE,
  parallelize = FALSE,
  allPars = FALSE,
  n.iter = 1000,
  n.burn = NULL,
  n.thin = 5,
  n.chain = 3,
  minSite = 1,
  maxSp = 9999
)
}
\arguments{
\item{filename}{the name of the datafile}

\item{inPath}{location where the \code{filename} can be found}

\item{outPath}{location where the results should be written. If \code{NULL} then results are returned rather than written.}

\item{useNimble}{option to bypass the model fitting in Nimble (just for testing the code)}

\item{inclPhenology}{should the model account for seasonal variation?}

\item{incl2ndTransect}{should the model include data from the second transect walk?}

\item{inclPanTrap}{should the model include pan trap data?}

\item{inclStateRE}{should there be a site-level random effect in the state model?}

\item{multiSp}{should the model be run for each species separately, or in a single model?}

\item{parallelize}{option to parallelize across MCMC chains}

\item{allPars}{if \code{TRUE} then all model parameters are monitored. If \code{FALSE}, just \code{mu.lambda} and \code{Trend}.}

\item{n.iter}{number of iterations for the Nimble model. Default is 1000.}

\item{n.burn}{number of iterations for the burn-in. If \code{NULL} (the default), then it will be set to \code{n.iter/2}.}

\item{n.thin}{thinning for the MCMC chains. Defaults to 5}

\item{n.chain}{number of MCMC chains. Defaults to 3}

\item{minSite}{the threshold minimum number of sites for a species to be considered for modelling}

\item{maxSp}{defines the maximum number of species to model. Species with numbers greater than this are ignored}
}
\value{
if \code{outpath} is NULL then a list comprising model output and metadata. Otherwise nothing
}
\description{
ProcessSimDatFile
}
\details{
Processes one file produced by Nacho's simulations. Extracts metadata, formats the data for the model, then runs the model. The file must be in RDS format.
}
