% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessSimFiles.R
\name{ProcessSimFiles}
\alias{ProcessSimFiles}
\title{ProcessSimFiles}
\usage{
ProcessSimFiles(
  inPath = ".",
  outPath = "output",
  collateOutput = TRUE,
  useNimble = FALSE,
  n.iter = 1000,
  inclPhenology = TRUE,
  inclPanTrap = TRUE,
  multiSp = FALSE,
  maxSp = 9999,
  maxFiles = 9999,
  parallelize = FALSE
)
}
\arguments{
\item{inPath}{location of folder containing many simulation runs}

\item{outPath}{location of folder to store results}

\item{collateOutput}{should the output be stored for collation all together (\code{TRUE}) or written as separate files (\code{FALSE})}

\item{useNimble}{option to bypass the model fitting in Nimble (just for testing the code)}

\item{n.iter}{number of iterations for the Nimble model}

\item{inclPhenology}{should the model account for seasonal variation?}

\item{inclPanTrap}{should the model include pan trap data?}

\item{multiSp}{should the model be run for species separately, or in a multispecies model?}

\item{maxSp}{defines the maximum number of species to model. Species with numbers greater than this are ignored}

\item{maxFiles}{defines the maximum number of files to work with. Additional files are ignored}

\item{parallelize}{option to parallelize across MCMC chains}
}
\value{
Nothing
from reshape2 @import dcast
}
\description{
ProcessSimFiles
}
\details{
Takes the name of a folder and processes them in a large batch
}
