% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessSimFolder.R
\name{ProcessSimFolder}
\alias{ProcessSimFolder}
\title{ProcessSimFolder}
\usage{
ProcessSimFolder(
  inPath = ".",
  outPath = "output",
  collateOutput = FALSE,
  useNimble = TRUE,
  inclPhenology = TRUE,
  inclPanTrap = TRUE,
  multiSp = FALSE,
  parallelize = FALSE,
  n.iter = 1000,
  n.burn = NULL,
  n.thin = 5,
  n.chain = 3,
  maxSp = 9999,
  maxFiles = 9999
)
}
\arguments{
\item{inPath}{location of folder containing many simulation runs}

\item{outPath}{location of folder to store results}

\item{collateOutput}{should the output be stored for collation all together (\code{TRUE}) or written as separate files (\code{FALSE})}

\item{useNimble}{option to bypass the model fitting in Nimble (just for testing the code)}

\item{inclPhenology}{should the model account for seasonal variation?}

\item{inclPanTrap}{should the model include pan trap data?}

\item{multiSp}{should the model be run for species separately, or in a multispecies model?}

\item{parallelize}{option to parallelize across MCMC chains}

\item{n.iter}{number of iterations for the Nimble model. Default is 1000.}

\item{n.burn}{number of iterations for the burn-in. If \code{NULL} (the default), then it will be set to \code{n.iter/2}.}

\item{n.thin}{thinning for the MCMC chains. Defaults to 5}

\item{n.chain}{number of MCMC chains. Defaults to 3}

\item{maxSp}{defines the maximum number of species to model. Species with numbers greater than this are ignored}

\item{maxFiles}{defines the maximum number of files to work with. Additional files are ignored}
}
\value{
Nothing
}
\description{
ProcessSimFolder
}
\details{
Takes the name of a folder and processes them in a large batch. The folder can be a zipped archive.
The files must be RDS format.
}
