% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{runModel}
\alias{runModel}
\title{runModel}
\usage{
runModel(
  dataConstants,
  obsData,
  dataSumm,
  useNimble = TRUE,
  inclPanTrap = TRUE,
  incl2ndTransect = TRUE,
  inclPhenology = TRUE,
  inclStateRE = FALSE,
  multiSp = FALSE,
  parallelize = FALSE,
  allPars = FALSE,
  n.iter = 1000,
  n.burn = NULL,
  n.thin = 5,
  n.chain = 3,
  maxSp = 9999
)
}
\arguments{
\item{dataConstants}{dataframe produced by ProcessSimDatFile()}

\item{obsData}{dataframe produced by ProcessSimDatFile()}

\item{useNimble}{option to bypass the model fitting in Nimble (just for testing the code)}

\item{inclPanTrap}{should the model include pan trap data?}

\item{incl2ndTransect}{should the model include data from the second transect walk?}

\item{inclPhenology}{should the model account for seasonal variation?}

\item{inclStateRE}{should there be a site-level random effect in the state model?}

\item{multiSp}{should the model be run as a multispecies model, or many single-species models?}

\item{parallelize}{should the chains be run as separate processes on different cores?}

\item{allPars}{if \code{TRUE} then all model parameters are monitored. If \code{FALSE}, just \code{lam.0bda} and \code{Trend}.}

\item{n.iter}{number of iterations for the Nimble model. Default is 1000.}

\item{n.burn}{number of iterations for the burn-in. If \code{NULL} (the default), then it will be set to \code{n.iter/2}.}

\item{n.thin}{thinning for the MCMC chains. Defaults to 5}

\item{n.chain}{number of MCMC chains. Defaults to 3}

\item{maxSp}{maximum number of species to be modelled}

\item{dataSumm$stats}{dataframe produced by ProcessSimDatFile()}
}
\value{
a set of year effects
}
\description{
runModel
}
\details{
Runs the model.
}
