% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{formatData}
\alias{formatData}
\title{formatData}
\usage{
formatData(inData, minSite = 1, incl2ndTransect = TRUE, inclPanTrap = TRUE)
}
\arguments{
\item{inData}{A dataset produced by the simulations}

\item{minSite}{the threshold minimum number of sites for a species to be considered for modelling}

\item{incl2ndTransect}{should the model include data from the second transect walk?}

\item{inclPanTrap}{should the model include pan trap data?}
}
\value{
list of two data frames
}
\description{
formatData
}
\details{
Formats data ready for Nimble
}
