% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessSimDatFile.R
\name{ProcessSimDatFile}
\alias{ProcessSimDatFile}
\title{ProcessSimDatFile}
\usage{
ProcessSimDatFile(
  filename,
  inPath = ".",
  outPath = NULL,
  useNimble = TRUE,
  n.iter = 1000,
  inclPhenology = TRUE,
  inclPanTrap = TRUE,
  maxSp = 9999,
  multiSp = FALSE,
  parallelize = FALSE
)
}
\arguments{
\item{filename}{the name of the datafile}

\item{inPath}{location where the \code{filename} can be found}

\item{outPath}{location where the results should be written. If \code{NULL} then results are returned rather than written.}

\item{useNimble}{option to bypass the model fitting in Nimble (just for testing the code)}

\item{n.iter}{number of iterations for the Nimble model}

\item{inclPhenology}{should the model account for seasonal variation?}

\item{inclPanTrap}{should the model include pan trap data?}

\item{maxSp}{defines the maximum number of species to model. Species with numbers greater than this are ignored}

\item{multiSp}{should the model be run for each species separately, or in a single model?}

\item{parallelize}{option to parallelize across MCMC chains}
}
\value{
if \code{outpath} is NULL then a list comprising model output and metadata. Otherwise nothing
}
\description{
ProcessSimDatFile
}
\details{
Processes one file produced by Nacho's simulations. Extracts metadata, formats the data for the model, then runs the model
}
