% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{runModel}
\alias{runModel}
\title{runModel}
\usage{
runModel(
  dataConstants,
  obsData,
  dataSumm,
  useNimble = TRUE,
  n.iter = 1000,
  inclPhenology = TRUE,
  inclPanTrap = TRUE,
  maxSp = 9999,
  parallelize = FALSE
)
}
\arguments{
\item{dataConstants}{dataframe produced by ProcessSimDatFile()}

\item{obsData}{dataframe produced by ProcessSimDatFile()}

\item{dataSumm}{dataframe produced by ProcessSimDatFile()}

\item{useNimble}{option to bypass the model fitting in Nimble (just for testing the code)}

\item{n.iter}{number of iterations for the Nimble model}

\item{inclPhenology}{should the model account for seasonal variation?}

\item{inclPanTrap}{should the model include pan trap data?}

\item{maxSp}{maximum number of species. If set at 1, the model is run sequentially for every species in the dataset}

\item{parallelize}{should the chains be run as separate processes on different cores?}
}
\value{
a set of year effects
}
\description{
runModel
}
\details{
Runs the model.
}
